requirements.txt
numpy
pandas
matplotlib
scikit-learn
python-dotenv
mysql-connector-python
pymysql
#-e .